Create README.md
The objective of vegetables Recognition using image processing is to design a incremental model to recognize the vegetables based on size, shape and colour of the vegetables ignoring external features like environment, noise and background. This just focuses on the image of a particular vegetable and identify the vegetables.
I made a project on machine learning, that is vegetable detection. By using this project a person can recognize which vegetable is that.